Correct nyaa.si
Old code is point to nyaa.si and use torrent file.
I have changed the info link to sukebei.nyaa.si and change the torrent url file to magnet.
My code maynot clean and I thought your code is working like a charm so it no need to modify that much.

Thanks a lot for this great script.